Make it work with multiple sites
The idea is to have a single node instance handling multiple sites
where each site has it's own directory structure.
This is especially useful for development where you don't want to
install and setup a full web server (like Apache or Nginx).

In this setup, you want to pass the sites to chokidar so that when
files change, the page is hot reloaded.

Suppose we have two sites:
- one
- two
When we change the file one/app.js, the emitted event will receive
the full path, i.e. one/app.js, while the hot reloader expect to
have a path without the prefix (in this case, one/).

This commit remove the prefix from the path so that the hot reloader
works as expected.

I've also created a small example to show how this works:
https://github.com/oligot/jspm-hot-reloader-multiple