Add capability to reload modules manually
Often when hot reloading I need to refresh the page (due to errors, for example). A function like this could help with that. 

A user can assign a key mapping to it in the browser, and reload just their modules excluding jspm_packages, or call it based on any other scenario he chooses.